requirements.txt
Flask==0.12.2
flask-admin==1.5.0
flask-security==3.0.0
flask-sqlalchemy==2.3.2
flask-mysqldb==0.2.0
InstagramApi==1.0.2
